Refs #21188: Move type support to dedicated directory
