+++ examples/cpp/request_reply/ServerApp.hpp
@@ -42,7 +42,7 @@
 
 #include "app_utils.hpp"
 #include "Application.hpp"
-#include "CalculatorPubSubTypes.hpp"
+#include "types/CalculatorPubSubTypes.hpp"
 
 namespace eprosima {
 namespace fastdds {
