+++ examples/cpp/request_reply/types/CalculatorCdrAux.hpp
