+++ examples/cpp/request_reply/types/CalculatorTypeObjectSupport.cxx
